requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
PyYAML==5.1.2
pandas==0.24.2
numpy==1.17.0
ortools==7.3.7083
XlsxWriter==1.2.0
xlrd==1.2.0
google-cloud-storage==1.19.0
gunicorn==19.9.0
